move create_action_fingerprints to trackers.py
